refactor(custom transformation): create abstract StringTransformation class.
