+++ src/main/java/com/conveyal/datatools/manager/models/transform/AddCustomFileFromStringTransformation.java
@@ -1,7 +1,7 @@
 package com.conveyal.datatools.manager.models.transform;
 import com.conveyal.datatools.common.status.MonitorableJob;
 
-public class AddCustomFileFromStringTransformation extends ReplaceFileFromStringTransformation {
+public class AddCustomFileFromStringTransformation extends StringTransformation {
 
     @Override
     public void validateTableName(MonitorableJob.Status status) {
